api/requirements/base.txt
Django==4.2.3
psycopg==3.1.9
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.1
pre-commit==3.2.2
# jwt
djangorestframework-simplejwt==5.2.2
django-model-utils==4.3.1
# env
django-environ==0.10.0

# celery
celery==5.3.5
django-celery-beat==2.5.0  # unlocks Django's database to store periodic tasks to be run by celery
django-celery-results==2.5.1  # enables to store the results of any given celery task in the Django database

# redis
redis==5.0.1  # enables Python to communicate with a Redis server.

# sendgrid to send mail
sendgrid==6.10.0
psycopg2==2.9.9
gunicorn==21.2.0

django-cte-trees==0.9.2

# Gevent
gevent==24.2.1
psycogreen==1.0.2

# Faker
faker==24.11.0

# Slug
python-slugify==8.0.4

# cors header
django-cors-headers==4.4.0

# channels
channels==4.1.0

#channels-redis
channels-redis==4.2.0

# daphne
daphne==4.1.2

whitenoise==6.8.2
